backend/requirements.txt
pywebview[qt]
fastapi
uvicorn[standard]
pyserial
pydantic
pydantic-settings
pyinstaller
psutil

